+++ src/IoTSettings.ts
@@ -1,29 +1,61 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
+import * as sdk from 'vscode-iot-device-cube-sdk';
+
 import {ConfigHandler} from './configHandler';
 import {PickWithData} from './Models/Interfaces/UI';
 
-export class IoTWorkbenchSettings {
-  private _workbenchPath: string;
 
-  constructor() {
-    const homeDir = os.homedir();
-    const platform = os.platform();
+export class IoTWorkbenchSettings {
+  private _workbenchPath = '';
+
+  private constructor() {}
+
+  static async createAsync() {
+    const iotWorkbenchSettings = new IoTWorkbenchSettings();
+
+    iotWorkbenchSettings._workbenchPath =
+        await IoTWorkbenchSettings.getWorkbenchPath();
+
+    return iotWorkbenchSettings;
+  }
+
+  static async getPlatform(): Promise<string> {
+    const localOs = sdk.Utility.require('os') as typeof import('os');
+    const getPlatform = await localOs.platform;
+    const platform = await getPlatform();
+    return platform;
+  }
+
+  static async getOs(): Promise<string> {
+    const localOs = sdk.Utility.require('os') as typeof import('os');
+    const getHomeDir = await localOs.homedir;
+    const homeDir = await getHomeDir();
+    return homeDir;
+  }
+
+  static async getWorkbenchPath(): Promise<string> {
+    const localOs = sdk.Utility.require('os') as typeof import('os');
+    const getHomeDir = await localOs.homedir;
+    const homeDir = await getHomeDir();
+    const getPlatform = await localOs.platform;
+    const platform = await getPlatform();
+
+    let _workbenchPath = '';
     if (platform === 'win32') {
-      this._workbenchPath =
-          path.join(homeDir, 'Documents', 'IoTWorkbenchProjects');
+      _workbenchPath = path.join(homeDir, 'Documents', 'IoTWorkbenchProjects');
     } else if (platform === 'linux') {
-      this._workbenchPath = path.join(homeDir, 'IoTWorkbenchProjects');
+      _workbenchPath = path.join(homeDir, 'IoTWorkbenchProjects');
     } else if (platform === 'darwin') {
-      this._workbenchPath =
-          path.join(homeDir, 'Documents', 'IoTWorkbenchProjects');
+      _workbenchPath = path.join(homeDir, 'Documents', 'IoTWorkbenchProjects');
     } else {
-      this._workbenchPath = '/IoTWorkbenchProjects';
+      _workbenchPath = '/IoTWorkbenchProjects';
     }
+
+    return _workbenchPath;
   }
 
   async workbenchPath() {