+++ src/Models/AZ3166Device.ts
@@ -21,12 +21,12 @@ import {delay, getRegistryValues} from '../utils';
 
 import {ArduinoDeviceBase} from './ArduinoDeviceBase';
 import {DeviceType} from './Interfaces/Device';
+import {TemplateFileInfo} from './Interfaces/ProjectTemplate';
 
 const impor = require('impor')(__dirname);
 const forEach = impor('lodash.foreach') as typeof import('lodash.foreach');
 const trimStart =
     impor('lodash.trimstart') as typeof import('lodash.trimstart');
-const filter = impor('lodash.filter') as typeof import('lodash.filter');
 
 interface SerialPortInfo {
   comName: string;
@@ -36,8 +36,6 @@ interface SerialPortInfo {
 }
 
 const constants = {
-  boardInfo: 'AZ3166:stm32f4:MXCHIP_AZ3166',
-  uploadMethod: 'upload_method=OpenOCDMethod',
   outputPath: './.build',
   platformLocalFileName: 'platform.local.txt',
   cExtraFlag: 'compiler.c.extra_flags=-DCORRELATIONID="',
@@ -49,10 +47,6 @@ const constants = {
 enum configDeviceOptions {
   ConnectionString,
   UDS
-}
-
-async function cmd(command: string) {
-  exec(command, Promise.resolve);
 }
 
 export class AZ3166Device extends ArduinoDeviceBase {
@@ -74,7 +68,7 @@ export class AZ3166Device extends ArduinoDeviceBase {
     return this.componentId;
   }
 
-  private sketchName = '';
+  private templateFiles: TemplateFileInfo[] = [];
   private static _boardId = 'devkit';
 
   static get boardId() {
@@ -83,19 +77,19 @@ export class AZ3166Device extends ArduinoDeviceBase {
 
   constructor(
       context: vscode.ExtensionContext, channel: vscode.OutputChannel,
-      devicePath: string, sketchName?: string) {
+      devicePath: string, templateFiles?: TemplateFileInfo[]) {
     super(context, devicePath, DeviceType.MXChip_AZ3166);
     this.channel = channel;
     this.componentId = Guid.create().toString();
-    if (sketchName) {
-      this.sketchName = sketchName;
+    if (templateFiles) {
+      this.templateFiles = templateFiles;
     }
   }
 
   name = 'AZ3166';
 
   get board() {
-    const boardProvider = new BoardProvider(this.extensionContext);
+    const boardProvider = new BoardProvider(this.boardFolderPath);
     const az3166 = boardProvider.find({id: AZ3166Device._boardId});
     return az3166;
   }
@@ -129,30 +123,11 @@ export class AZ3166Device extends ArduinoDeviceBase {
       throw new Error('Unable to find the board in the config file.');
     }
 
-    this.generateCppPropertiesFile(this.board);
-
     return true;
   }
 
   async create(): Promise<boolean> {
-    if (!this.sketchName) {
-      throw new Error('No sketch file found.');
-    }
-    const deviceFolderPath = this.deviceFolder;
-
-    if (!fs.existsSync(deviceFolderPath)) {
-      throw new Error('Unable to find the device folder inside the project.');
-    }
-    if (!this.board) {
-      throw new Error('Unable to find the board in the config file.');
-    }
-
-    this.generateCommonFiles();
-    this.generateCppPropertiesFile(this.board);
-    await this.generateSketchFile(
-        this.sketchName, this.board, constants.boardInfo,
-        constants.uploadMethod);
-    return true;
+    return this.createCore(this.board, this.templateFiles);
   }
 
   async preCompileAction(): Promise<boolean> {