+++ src/Models/Esp32Device.ts
@@ -10,20 +10,17 @@ import * as vscode from 'vscode';
 
 import {BoardProvider} from '../boardProvider';
 import {ConfigHandler} from '../configHandler';
-import {ConfigKey} from '../constants';
-
+import {ConfigKey, FileNames, ScaffoldType} from '../constants';
+import {FileUtility} from '../FileUtility';
+import {IoTWorkbenchSettings} from '../IoTSettings';
 
 import {ArduinoDeviceBase} from './ArduinoDeviceBase';
 import {DeviceType} from './Interfaces/Device';
-
-const constants = {
-  defaultBoardInfo: 'esp32:esp32:m5stack-core-esp32',
-  defaultBoardConfig:
-      'FlashMode=qio,FlashFreq=80,UploadSpeed=921600,DebugLevel=none'
-};
+import {TemplateFileInfo} from './Interfaces/ProjectTemplate';
+import {IoTWorkbenchProjectBase} from './IoTWorkbenchProjectBase';
 
 export class Esp32Device extends ArduinoDeviceBase {
-  private sketchFileTemplateName = '';
+  private templateFiles: TemplateFileInfo[] = [];
   private static _boardId = 'esp32';
   private channel: vscode.OutputChannel;
 
@@ -37,7 +34,7 @@ export class Esp32Device extends ArduinoDeviceBase {
   }
 
   get board() {
-    const boardProvider = new BoardProvider(this.extensionContext);
+    const boardProvider = new BoardProvider(this.boardFolderPath);
     const esp32 = boardProvider.find({id: Esp32Device._boardId});
     return esp32;
   }
@@ -70,12 +67,12 @@ export class Esp32Device extends ArduinoDeviceBase {
 
   constructor(
       context: vscode.ExtensionContext, channel: vscode.OutputChannel,
-      devicePath: string, sketchFileTemplateName?: string) {
+      devicePath: string, templateFiles?: TemplateFileInfo[]) {
     super(context, devicePath, DeviceType.IoT_Button);
     this.channel = channel;
     this.componentId = Guid.create().toString();
-    if (sketchFileTemplateName) {
-      this.sketchFileTemplateName = sketchFileTemplateName;
+    if (templateFiles) {
+      this.templateFiles = templateFiles;
     }
   }
 
@@ -86,7 +83,8 @@ export class Esp32Device extends ArduinoDeviceBase {
   async load(): Promise<boolean> {
     const deviceFolderPath = this.deviceFolder;
 
-    if (!fs.existsSync(deviceFolderPath)) {
+    if (!await FileUtility.directoryExists(
+            ScaffoldType.Workspace, deviceFolderPath)) {
       throw new Error('Unable to find the device folder inside the project.');
     }
 
@@ -94,31 +92,12 @@ export class Esp32Device extends ArduinoDeviceBase {
       throw new Error('Unable to find the board in the config file.');
     }
 
-    this.generateCppPropertiesFile(this.board);
     return true;
   }
 
   async create(): Promise<boolean> {
-    if (!this.sketchFileTemplateName) {
-      throw new Error('No sketch file found.');
-    }
-    const deviceFolderPath = this.deviceFolder;
-
-    if (!fs.existsSync(deviceFolderPath)) {
-      throw new Error('Unable to find the device folder inside the project.');
-    }
-    if (!this.board) {
-      throw new Error('Unable to find the board in the config file.');
-    }
-
-    this.generateCommonFiles();
-    this.generateCppPropertiesFile(this.board);
-    await this.generateSketchFile(
-        this.sketchFileTemplateName, this.board, constants.defaultBoardInfo,
-        constants.defaultBoardConfig);
-    return true;
+    return this.createCore(this.board, this.templateFiles);
   }
-
 
   async configDeviceSettings(): Promise<boolean> {
     const configSelectionItems: vscode.QuickPickItem[] = [